Rename dummy folder
The folder Dummy has been renamed to DummyUtils
in order to fit the utilities specifications.
